Week_10_Lab/requirements.txt
streamlit
bcrypt
pandas
numpy
plotly
openai
python-dotenv
